+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/ITEEnchantingTableMenu.java
@@ -122,7 +122,7 @@ public class ITEEnchantingTableMenu extends AbstractContainerMenu {
                 if (stack.getTag().contains(EnchantingFeature.EMPOWERED_ITEM))
                     p *= 1.25f;
             }
-            float maxCost = (EnchantmentsFeature.getEnchantmentValue(stack) + EnchantingFeature.getFirstCurseCost(stack)) * p * (enchantingPower / 15f) + 4;
+            float maxCost = (EnchantmentsFeature.getEnchantmentValue(stack) + EnchantingFeature.getCurseCost(stack)) * p * (enchantingPower / 15f) + 4;
             this.maxCost.set((int) (maxCost * 100f));
         }
         this.broadcastChanges();