+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/GravityDefying.java
@@ -1,8 +1,12 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraftforge.common.ForgeMod;
@@ -11,7 +15,7 @@ import net.minecraftforge.event.entity.living.LivingFallEvent;
 
 import java.util.UUID;
 
-public class GravityDefying extends Enchantment {
+public class GravityDefying extends Enchantment implements IAttributeEnchantment, IEnchantmentTooltip {
     public static final UUID GRAVITY_MODIFIER_UUID = UUID.fromString("7567b3ad-a4c6-4700-8bf0-cd8c4356c155");
     public GravityDefying() {
         super(Rarity.RARE, EnchantmentCategory.ARMOR_FEET, new EquipmentSlot[] {EquipmentSlot.FEET});
@@ -36,21 +40,21 @@ public class GravityDefying extends Enchantment {
         return lvl * 0.6f - ((lvl - 1) * 0.45f);
     }
 
-    public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
-        if (event.getSlotType() != EquipmentSlot.FEET)
-            return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.GRAVITY_DEFYING.get());
-        if (lvl <= 0)
-            return;
-
-        event.addModifier(ForgeMod.ENTITY_GRAVITY.get(), new AttributeModifier(GRAVITY_MODIFIER_UUID, "Gravity Defying enchantment", -getGravityReduction(lvl), AttributeModifier.Operation.MULTIPLY_BASE));
-    }
-
     public static void applyFallDamageReduction(LivingFallEvent event) {
         int lvl = event.getEntity().getItemBySlot(EquipmentSlot.FEET).getEnchantmentLevel(NewEnchantmentsFeature.GRAVITY_DEFYING.get());
         if (lvl <= 0)
             return;
 
         event.setDistance((event.getDistance() - lvl * 4));
+    }
+
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
+        event.addModifier(ForgeMod.ENTITY_GRAVITY.get(), new AttributeModifier(GRAVITY_MODIFIER_UUID, "Gravity Defying enchantment", -getGravityReduction(enchantmentLvl), AttributeModifier.Operation.MULTIPLY_BASE));
+    }
+
+    @Override
+    public Component getTooltip(ItemStack itemStack, int lvl) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.DARK_PURPLE);
     }
 }