+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Reach.java
@@ -1,8 +1,11 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage.BonusDamageEnchantment;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.item.ItemStack;
@@ -12,7 +15,7 @@ import net.minecraftforge.event.ItemAttributeModifierEvent;
 
 import java.util.UUID;
 
-public class Reach extends Enchantment/* implements IEnchantmentTooltip */{
+public class Reach extends Enchantment implements IAttributeEnchantment, IEnchantmentTooltip {
 
     public static final UUID REACH_MODIFIER_UUID = UUID.fromString("75f2bcd4-4f76-43b8-b75c-0c2a825cec8a");
 
@@ -40,19 +43,16 @@ public class Reach extends Enchantment/* implements IEnchantmentTooltip */{
         return super.canApplyAtEnchantingTable(stack) || stack.is(BonusDamageEnchantment.ACCEPTS_DAMAGE_ENCHANTMENTS);
     }
 
-    public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
         if (event.getSlotType() != EquipmentSlot.MAINHAND)
             return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.REACH.get());
-        if (lvl == 0)
-            return;
-
-        event.addModifier(ForgeMod.ENTITY_REACH.get(), new AttributeModifier(REACH_MODIFIER_UUID, "Reach Enchantment Modifier", 0.2f * lvl, AttributeModifier.Operation.MULTIPLY_BASE));
-        event.addModifier(ForgeMod.BLOCK_REACH.get(), new AttributeModifier(REACH_MODIFIER_UUID, "Reach Enchantment Modifier", 0.2f * lvl, AttributeModifier.Operation.MULTIPLY_BASE));
+        event.addModifier(ForgeMod.ENTITY_REACH.get(), new AttributeModifier(REACH_MODIFIER_UUID, "Reach Enchantment Modifier", 0.2f * enchantmentLvl, AttributeModifier.Operation.MULTIPLY_BASE));
+        event.addModifier(ForgeMod.BLOCK_REACH.get(), new AttributeModifier(REACH_MODIFIER_UUID, "Reach Enchantment Modifier", 0.2f * enchantmentLvl, AttributeModifier.Operation.MULTIPLY_BASE));
     }
 
-    /*@Override
+    @Override
     public Component getTooltip(ItemStack stack, int lvl) {
         return Component.translatable(this.getDescriptionId() + ".tooltip", IguanaTweaksReborn.ONE_DECIMAL_FORMATTER.format(0.2f * lvl * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
-    }*/
+    }
 }