+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/StepUp.java
@@ -1,8 +1,11 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraftforge.common.ForgeMod;
@@ -10,7 +13,7 @@ import net.minecraftforge.event.ItemAttributeModifierEvent;
 
 import java.util.UUID;
 
-public class StepUp extends Enchantment {
+public class StepUp extends Enchantment implements IAttributeEnchantment, IEnchantmentTooltip {
 
     public static final UUID MODIFIER_UUID = UUID.fromString("74e97c20-6a62-482f-b909-e709087b066a");
 
@@ -28,14 +31,13 @@ public class StepUp extends Enchantment {
         return 40;
     }
 
-    public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
-        if (event.getSlotType() != EquipmentSlot.LEGS)
-            return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.STEP_UP.get());
-        if (lvl == 0)
-            return;
-
-        event.addModifier(ForgeMod.STEP_HEIGHT_ADDITION.get(), new AttributeModifier(MODIFIER_UUID, "Step Up Enchantment Modifier", 0.5d * lvl, AttributeModifier.Operation.ADDITION));
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
+        event.addModifier(ForgeMod.STEP_HEIGHT_ADDITION.get(), new AttributeModifier(MODIFIER_UUID, "Step Up Enchantment Modifier", 0.5d * enchantmentLvl, AttributeModifier.Operation.ADDITION));
     }
 
+    @Override
+    public Component getTooltip(ItemStack itemStack, int i) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.DARK_PURPLE);
+    }
 }