+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/SwiftStrike.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage.BonusDamageEnchantment;
@@ -14,7 +13,7 @@ import net.minecraftforge.event.ItemAttributeModifierEvent;
 
 import java.util.UUID;
 
-public class SwiftStrike extends BonusDamageEnchantment {
+public class SwiftStrike extends BonusDamageEnchantment implements IAttributeEnchantment {
     public static final UUID BONUS_ATTACK_SPEED_UUID = UUID.fromString("7b0cb3a4-7a7c-4908-be8d-aadd523690d7");
     public SwiftStrike() {
         super(Rarity.UNCOMMON, EnchantmentsFeature.WEAPONS_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
@@ -29,18 +28,15 @@ public class SwiftStrike extends BonusDamageEnchantment {
         return 0.1f;
     }
 
-    public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
-        if (event.getSlotType() != EquipmentSlot.MAINHAND)
-            return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.SWIFT_STRIKE.get());
-        if (lvl == 0)
-            return;
-
-        event.addModifier(Attributes.ATTACK_SPEED, new AttributeModifier(BONUS_ATTACK_SPEED_UUID, "Rhythmic Swing enchantment", getAttackSpeedBonusPerLevel() * lvl, AttributeModifier.Operation.MULTIPLY_BASE));
-    }
-
     @Override
     public Component getTooltip(ItemStack stack, int lvl) {
         return Component.translatable(this.getDescriptionId() + ".tooltip", IguanaTweaksReborn.ONE_DECIMAL_FORMATTER.format(getAttackSpeedBonusPerLevel() * lvl * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
+    }
+
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
+        if (event.getSlotType() != EquipmentSlot.MAINHAND)
+            return;
+        event.addModifier(Attributes.ATTACK_SPEED, new AttributeModifier(BONUS_ATTACK_SPEED_UUID, "Rhythmic Swing enchantment", getAttackSpeedBonusPerLevel() * enchantmentLvl, AttributeModifier.Operation.MULTIPLY_BASE));
     }
 }