+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Zippy.java
@@ -1,16 +1,20 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.insanelib.InsaneLib;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraftforge.event.ItemAttributeModifierEvent;
 
 import java.util.UUID;
 
-public class Zippy extends Enchantment {
+public class Zippy extends Enchantment implements IAttributeEnchantment, IEnchantmentTooltip {
 
     public static final UUID MODIFIER_UUID = UUID.fromString("74e97c20-6a62-482f-b909-e709087b066a");
 
@@ -33,14 +37,13 @@ public class Zippy extends Enchantment {
         return this.getMinCost(level) + 22;
     }
 
-    public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
-        if (event.getSlotType() != EquipmentSlot.LEGS)
-            return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.ZIPPY.get());
-        if (lvl == 0)
-            return;
-
-        event.addModifier(Attributes.MOVEMENT_SPEED, new AttributeModifier(MODIFIER_UUID, "Zippy Enchantment Modifier", 0.1d * lvl, AttributeModifier.Operation.MULTIPLY_BASE));
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
+        event.addModifier(Attributes.MOVEMENT_SPEED, new AttributeModifier(MODIFIER_UUID, "Zippy Enchantment Modifier", 0.1d * enchantmentLvl, AttributeModifier.Operation.MULTIPLY_BASE));
     }
 
+    @Override
+    public Component getTooltip(ItemStack itemStack, int lvl) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip", InsaneLib.ONE_DECIMAL_FORMATTER.format(10f + lvl)).withStyle(ChatFormatting.DARK_PURPLE);
+    }
 }