+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/curse/CurseOfExperience.java
@@ -18,7 +18,7 @@ import net.minecraft.world.item.enchantment.EnchantmentCategory;
 public class CurseOfExperience extends Enchantment implements IEnchantmentTooltip {
     public static ResourceKey<DamageType> DAMAGE_TYPE = ResourceKey.create(Registries.DAMAGE_TYPE, new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "curse_of_experience"));
     public CurseOfExperience() {
-        super(Rarity.VERY_RARE, EnchantmentCategory.VANISHABLE, EquipmentSlot.values());
+        super(Rarity.VERY_RARE, EnchantmentCategory.DIGGER, EquipmentSlot.values());
     }
 
     @Override