+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/curse/CurseOfInefficiency.java
@@ -1,0 +1,38 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.curse;
+
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class CurseOfInefficiency extends Enchantment implements IEnchantmentTooltip {
+    public CurseOfInefficiency() {
+        super(Rarity.UNCOMMON, EnchantmentCategory.DIGGER, EquipmentSlot.values());
+    }
+
+    @Override
+    public boolean isTreasureOnly() {
+        return true;
+    }
+
+    @Override
+    public boolean isCurse() {
+        return true;
+    }
+
+    public int getMinCost(int pEnchantmentLevel) {
+        return 25;
+    }
+
+    public int getMaxCost(int pEnchantmentLevel) {
+        return 50;
+    }
+
+    @Override
+    public Component getTooltip(ItemStack itemStack, int i) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.RED);
+    }
+}